pyproject.toml
[project]
name = "priorauth-agent"
version = "0.0.1"
description = "Prior Authorization A2A agent for lumbar MRI — Agents Assemble hackathon submission."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Kevin Shine George" },
    { name = "Sanjit Saji" },
]

# Top-level repo is a uv workspace with three members: shared, mcp_server, a2a_agent.
# Each member has its own pyproject.toml for isolated deps.
# Shared dependencies used across all members are declared here.
dependencies = [
    "pydantic>=2.7,<3",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "google-generativeai>=0.7",
]

[dependency-groups]
dev = [
    "ruff>=0.5",
    "mypy>=1.11",
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "types-requests",
]

[tool.uv.workspace]
members = ["shared", "mcp_server", "a2a_agent"]

[tool.uv.sources]
shared = { workspace = true }

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["shared", "mcp_server", "a2a_agent", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "C4",  # comprehensions
    "SIM", # simplify
    "PL",  # pylint subset
    "RUF", # ruff-specific
]
ignore = [
    "PLR0913",  # too-many-arguments — Pydantic models often need many
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR2004"]  # magic values in tests are fine

# Upstream reference code (copied from prompt-opinion/po-adk-python). We sync
# this folder periodically; do not fight its upstream style. See a2a_agent/REFERENCE.md.
"a2a_agent/po_base/**" = [
    "E501",     # aligned dict formatting in logs/hooks
    "E402",     # module import after class def (intentional in app_factory.py)
    "I001",     # upstream import order preserved for easier diffs
    "PLR0912",  # middleware.dispatch branches by design
    "PLR0915",  # middleware.dispatch statements by design
    "PLC0415",  # deferred starlette import inside request scope
    "RUF012",   # upstream class-level ANSI color dict
    "B008",     # default args in upstream factory
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

# Upstream reference code (copied from prompt-opinion/po-adk-python). Mirrors
# the ruff per-file-ignore above; our own code stays strictly typechecked.
# See a2a_agent/REFERENCE.md.
[[tool.mypy.overrides]]
module = "a2a_agent.po_base.*"
ignore_errors = true

# Upstream libraries without bundled type stubs.
[[tool.mypy.overrides]]
module = ["a2a.*", "google.adk.*", "dotenv", "mcp.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "llm: tests that call a real LLM (can be skipped for fast CI)",
    "integration: end-to-end tests that require both services running",
]
